# SwEng . major compat fixes for interoperation with wxw. . add crashlog_sender (modified version of debgrpt wxw sample). compiles but untested, will be changed heavily further improvements: . wposix no longer declares CRT functions. instead, use CRT headers.. but don't allow their (incorrect/non-POSIX) declarations to propagate to users. . split up win.cpp into wstartup, wutil, winit . extern "C" -> EXTERN_C . move sdl.* to lib/external_libraries and add png.h+zlib.h (fixes for and include of the library)
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@5028 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/sysdep/win/wposix/wposix.cpp
@@ -24,6 +24,7 @@
 #include "wposix.h"
 
 #include "wposix_internal.h"
+#include "crt_posix.h"		// _getcwd
 
 
 long sysconf(int name)
@@ -104,7 +105,7 @@ long sysconf(int name)
 #endif
 char* getcwd(char* buf, size_t buf_size)
 {
-	return _getcwd(buf, buf_size);
+	return _getcwd(buf, (int)buf_size);
 }
 #ifdef REDEFINED_NEW
 # include "lib/mmgr.h"
